pyproject.toml
[project]
name = "satsuki"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "aiosqlite>=0.20.0",
    "demoji>=1.1.0",
    "discord-ext-menus>=1.1",
    "discord-py>=2.4.0",
    "discord-sentry-reporting>=2.0.4",
    "feedparser>=6.0.11",
    "html2text>=2024.2.26",
    "mojimoji>=0.0.13",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "sentry-sdk>=2.14.0",
    "sqlalchemy>=2.0.35",
]

[dependency-groups]
dev = [
    "ruff>=0.14.3",
]
